Create test suite and refactor towards using decimals

+++ Application/Core/Interfaces/OutputHandler.cs
@@ -2,6 +2,6 @@ namespace KoganCodingChallenge.Output
 {
     public interface OutputHandler
     {
-        public void OutputCubicWeight(float value);
+        public void OutputCubicWeight(decimal value);
     }
 }
